engine: Don't allow registration of invalid resources
It turns out that trying to use a resource with an unexported field in
the language will cause it to panic because of a likely mistake in the
upstream golang. In any case, since it was difficult to work around this
during runtime, and drop those fields, the sensible thing is to error on
resource registration to notify the programmer of their mistake.

Since the test cases run resource registration, this also causes them
all to be tested!